+++ Changelog.md
@@ -6,6 +6,7 @@
   * Wrapped kernel PCA (requires latestMultivariateStats.jl)
   * Added decision stump classifier, regressor
   * Added BKS label combiner
+  * Added support for generating kernels i.e. kernel trick 
 
 0.1.0 / 2017-09-24
 ==================